updated ec2 tests
